requirements.txt
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
Pillow==10.4.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
psycopg2-binary==2.9.9
psycopg==3.2.1
python-dotenv==1.0.1
black==24.8.0
